bot-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.0
sqlalchemy==2.0.36
asyncpg==0.30.0
langchain==0.3.7
langchain-openai==0.2.14
